Feat App + ImgSearch move state upwards
