+++ src/screens/ImgSearch/index.jsx
@@ -1,6 +1,5 @@
 // @flow
 import React, { Component } from 'react'
-import axios from 'axios'
 
 import SearchBar from '../../components/SearchBar'
 import ImagesDisplayed from '../../components/ImagesDisplayed'
@@ -17,86 +16,35 @@ interface ThumbImgSerialized {
 }
 
 type State = {
-  imgData: Array<Object>,
-  value: string,
+  data: Array<ThumbImgSerialized>,
+  search: string,
 }
 
-const emptyCollection: Array<ThumbImgData> = [
-  {
-    links: [
-      {
-        href: '',
-      },
-    ],
-    data: [
-      {
-        nasa_id: '',
-        title: '',
-      },
-    ],
-  },
-]
+type Props = {
+  data: Array<ThumbImgSerialized>,
+  getSearchResults: Function,
+}
 
-export default class ImgSearch extends Component<{}, State> {
-  collectRESTData: (data: Array<ThumbImgData>) => void
-  serializeData: (itemList: Array<ThumbImgData>) => Array<ThumbImgSerialized>
-  constructor() {
-    super()
+type DerivedSP = {
+  data: Array<ThumbImgSerialized>,
+}
+
+export default class ImgSearch extends Component<Props, State> {
+  constructor(props: Props) {
+    super(props)
     this.state = {
-      imgData: [{ key: 'KSC-99pp0498' }],
-      value: '',
+      data: [{ src: ' ', key: ' ', alt: ' ' }],
+      search: '',
     }
-    this.collectRESTData = this.collectRESTData.bind(this)
-    this.serializeData = this.serializeData.bind(this)
   }
-  collectRESTData(data: Array<ThumbImgData>) {
-    const itemList = data ? data : emptyCollection
-    const itemsSanitized = this.serializeData(itemList)
-    this.setState(() => {
-      return { imgData: itemsSanitized }
-    })
-  }
-  getSearchResults = (media_type: 'image' | 'audio') => {
-    const { value } = this.state
-    const queryPassed = value.length > 2 ? value : ''
-    return axios
-      .get(`https://images-api.nasa.gov/search?`, {
-        params: {
-          q: queryPassed,
-          media_type,
-        },
-      })
-      .then(response => {
-        if (response.status >= 200 && response.status < 400) {
-          const responseItemList = response.data.collection.items
-          this.collectRESTData(responseItemList.filter((_, i) => i < 20))
-        }
-        return
-      })
-      .catch(error => {
-        console.log(error)
-      })
+  static getDerivedStateFromProps = (
+    nextProps: DerivedSP,
+    prevState: DerivedSP
+  ) => {
+    return nextProps.data === prevState.data ? {} : { data: nextProps.data }
   }
   handleChange = (event: SyntheticInputEvent<HTMLInputElement>) => {
-    this.setState({ value: event.target.value })
-    this.getSearchResults('image')
-  }
-  serializeData(itemList: Array<ThumbImgData>): Array<ThumbImgSerialized> {
-    return itemList.map(items => {
-      let data: Array<{ nasa_id: string, title: string }> = items.data
-      let links: Array<{ href: string }> = items.links
-      const { nasa_id, title } = data[0]
-      const { href } = links[0]
-      return {
-        alt: title,
-        key: nasa_id,
-        src: href,
-      }
-    })
-  }
-  componentDidMount() {
-    this.setState({ value: '' })
-    this.getSearchResults('image')
+    this.props.getSearchResults('image', event.target.value)
   }
   render() {
     return (
@@ -104,7 +52,7 @@ export default class ImgSearch extends Component<{}, State> {
         <article className="search-bar-wrapper">
           <SearchBar callback={this.handleChange} />
         </article>
-        <ImagesDisplayed collection={this.state.imgData} />
+        <ImagesDisplayed collection={this.state.data} />
       </section>
     )
   }
